BIGTOP-23. Switching to pig 0.9.0 to allow for building with Java 6.
There's a Forrest 0.8/Java 6/Pig < 0.9.0 problem that leads to
hangs/out of heap space errors. This was fixed in Pig 0.9.0, so we'll
switch our pig version to that. This does end up also requiring
explicitly specifying the pig version on the ant calls in the Debian
and RPM builds, due to a snafu in the pig tarball where the version is
set to 0.9.1-SNAPSHOT, even though it's the 0.9.0 release
tarball. Some minor tweaks were needed to package.mk to support making
sure the needed variable was available in the Debian build.

git-svn-id: https://svn.apache.org/repos/asf/incubator/bigtop/trunk@1159446 13f79535-47bb-0310-9956-ffa450edef68