BIGTOP-1671. Fix handling symlinks when cloning git repos
This upgrades version of grgit. This version of
grgit uses a newer version of jgit which supports
native symbolic links. This allows git repos that
contain symbolic links to be properly cloned such
as the Hue repo.

For jgit to handling symbolic links,
the org.eclipse.jgit.java7 jar/bundle must be
added to the classpath.

Signed-off-by: Konstantin Boudnik <cos@apache.org>