BIGTOP-1838: Publish results of pig build to local ivy/maven repositories
The pig build does not execute the published target. This means that other pieces of the bigtop
ecosystem do not pick up the correct pig dependencies from the bigtop build. Instead, they pick
them up from the central maven repository (or they fail to find them if a bigtop has pig
version that is not publicly released).

This is a problem because builds like datafu pick up their pig dependency via ivy and they are not able to do so.

Signed-off-by: Konstantin Boudnik <cos@wandisco.com>