BIGTOP-1840: Add local resolver to datafu's ivy resolver chain
When building datafu there is a sed replacement done to make sure that datafu pulls in the same
version of pig that is in the BOM. However, the ivysettings for datafu do not actually resolve
to the local repository so datafu does not pick up the locally built pick for the Bigtop.
datafu instead grabs pig from the central repository. This is especially problematic when
dealing with a custom version of Pig as it cannot be resolved.

Signed-off-by: Konstantin Boudnik <cos@wandisco.com>