BIGTOP-2293. Add puppet recipes for qfs components
This change adds puppet recipes to the bigtop source code for qfs components:
qfs-metaserver, qfs-chunkserver, and qfs-client. The puppet recipes install the
necessary packages, create necessary files and directories, generate
configuration with the correct ports and paths, and finally start the services
so that they can talk to each other.

I have tested to make sure that the systems work properly by logging in manually
to the docker container and interacting with the chunkserver and metaserver
using the qfs client. In my next commit, I will add testing to the bigtop
codebase so that we don't need to use manual inspection to ensure that things
are working correctly.