colocate bundle applications
To save resources, colocate some of the applications when it makes sense.
Put the namenode and resourcemgr on the same unit; spark and zeppelin;
ganglia and rsyslog.

Update readme formatting and callout these colocated apps. Also tweak
the restricted network section and add a section for connecting to zepp.

Note, machine spec in the bundle yaml start with machine '1' instead of
'0'. This is because someone may deploy this on juju-1.x, where machine
'0' is the bootstrap node.