Zookeeper now does an automatic rolling restart after quorum change. (#46)
* Zookeeper now does an automatic rolling restart after quorum change.

Replaces user unfriendly manual restart process.

* Fixed an issue where things broke when you removed the Juju leader.

The leader still attempted to orchestrate things as it was shutting
down. We fix this by requiring zkpeer.joined on check_cluster, and add
an alternate version of the funciton that triggers on zkpeer.departed,
so the new leader picks things up.

* Fixed issue w/ unrelated relation data changing causing bugs during restart

We now update the queue based on which nodes have restarted, for the
current nonce, instead of trying to do the dance with states. This seems
to work well.