more sparkbench fixes
The last commit moved all sparkbench data locally, but when spark is in yarn
mode, the nodemanagers (acting as spark workers) can't access this local data.
Doh! Fix it by detecting our execution mode; if spark is in yarn mode, the
DATA_HDFS dir will need to point to an hdfs location, and the action will need
to generate/store input data for the benchmarks there.