surely the last round of sparkbench fixes
Building off the last commit which did the right thing for yarn vs local modes,
we also need to run the benchmark as a different user depending on the mode.

In yarn mode, the data files are stored in hdfs, so the benchmark needs to run
as a user guaranteed to have access to hdfs from the spark unit(s). That's
ubuntu.

In local mode, the benchmarks read/write intermediate files that need to be
accessible by the spark process, so make sure they are run as 'spark'.

Fix up directory permissions to ensure ubuntu+spark can do what they need
to do regardless of spark exe mode.