relax test charm definition
Testing a local charm change would require changing the test because it
explicitly pulled the charm-under-test from the store. Relax this to
use the charm name -- if run from a local charm repo with the charm
available, the test will use that. If run from somewhere without an
available local charm (like cwr), the test will combine the deployment
series + charm name and deploy from the store.