lint/readme tweaks and a rename
Rename zk lib with a 'bigtop_' prefix to be consistent with other bigtop
charms. Also tweak readme for consistency.

Noqa E402 on the actions. These aren't currently causing lint failures
because they don't end in .py, but we may as well noqa them because we
know we want E402 ignored (error when imports aren't at the top of a file).