Update bigtop-detect-javahome for variable typo
In line 56 and line 60, ${OPENJAVA8_HOME_CANDIDATES} can only be empty because there was a typo in defining variable OPENJAVA8_HOME__CANDIDATES in line 34.   There are two '_' between "HOME" and "CANDIDATES" in line 34. But There is only one '_'  between "HOME" and "CANDIDATES" when using the variable in line 56 and line 60. For other variable using one '_' as the separator instead of two '_', I suggest changing variable $OPENJAVA8_HOME__CANDIDATES(with two '_' between "HOME" and "CANDIDATES") in line 34 to $OPENJAVA8_HOME_CANDIDATES(with only one '_' between "HOME" and "CANDIDATES")