remove bigtop_version override for kafka
Since kafka is no longer available in the bigtop repos, we need to
install it from a CI repo. This should be a 1.2.1 repo, not master.
Remove the bigtop_version override from config.yaml to ensure
1.2.1 is used.

This will coincide with a PR to layer-apache-bigtop-base that handles
repo_url construction specifically for kafka.