BIGTOP-3040: Solr's default configsets doesn't work with smoke test cases
The default configsets used in solr is data_driven_schema_configs, which
doesn't work with solr smoke test, TestIndexing:testSearch. The reason is
in data_driven_schema_configs "name" is treated as StrField, not TextField,
so the words will not be tokenized, and lead to unexpected query result.
sample_techproducts_configs should be used.

This closes #364

Change-Id: I170c9530ed937410dcd99c975276c30777b24045
Signed-off-by: Jun He <junhe@apache.org>