BIGTOP-3440: disable flume solr sink to avoid build failures (#694)
The solr sink currently depends on kite, that it is a library not
developed anymore. There are currently build failures due to
the ua_parser library, that was available in a Twitter maven repository
not maintained anymore (since it throws HTTP 503s consistently).
Flume upstream moved the dependency to another repository,
https://github.com/apache/flume/commit/9e0bc5bf4865673e6989ba887d27f468991c1e27
that fixes the Twitter problem but that brings in another build failure
for the 'rome' dependency.

This commit removes the solr sink from the build list, resolving the
overall build problem for Flume.

(cherry picked from commit 8e1bf51585ef78a914640d7d3eba367a174d2a76)