BIGTOP-3455: add hive patch to fix HIVE-17368
In Kerberos-enabled environments, Hive < 2.4.x (still not released) suffers
the problem outlined in HIVE-17368. This patch is related to the following commit:

https://github.com/apache/hive/commit/b3a6e524a6cb17893cbbeb26a877b0196ba16e21

The commit is related to the backport of the patch for the Hive 2.x branch.

Tested on top of Bigtop 1.4 (the issue that I was experiencing with Oozie
got fixed) but I haven't tested it with Bigtop 1.5 (but I am reasonably sure
that it works fine).

Built the hive-pkg target on slave-debian10, no issue reported.