BIGTOP-3515 - Backport HIVE-19231
This patch is contained in the Hive 3.x branch, but not in the Hive 2.x one. It
leads to subtle errors when executing Hive via Python or systemd, and
having it fixed in the 1.5 branch would allow people to rebuild their
packages if needed (we could update the docs about it).