BIGTOP-3527: replace ${PWD} with ${CURDIR} in debian rules files
According to
https://lintian.debian.org/tags/debian-rules-calls-pwd.html,
CURDIR is the preferred way in a makefile (like rules) to refer to the
current directory. For some reason I got building failures when running
a Debian 10 bigtop docker imagei and trying to build the hadoop pkg,
that got solved only applying this fix.

Applied the same fix to all the debian rules files with ${PWD}.