BIGTOP-3609 - Update ca-certificates for CentOS 7
Due to https://access.redhat.com/articles/6338021 ant and  maven classes
fail to download due to TLS certificate expiration check issues.
Since the issue is related to CentOS 7 and earlier, it seems sensible
to force the upgrade of the ca-certificates package (that in its latest
versions contain a fix).

While at it:
1) Add a specific dependency to 'ca-certificates' for CentOS 7.
2) Require bigtop_toolchain::packages in the varios ant/maven/groovy
   classes, so that we'll be able to rely on packages being
   installed when running the Execs. It may not be strictly required
   now but it seems good for future code.
3) Refactor a bit the Execs to be more readable when puppet runs
   (easier to track what's happening).

Tested rebuilding the puppet image and the slave image.