BIGTOP-3613 - Upgrade hive 2.3.6 to log4j 2.17.0
This commit is based on the work done in BIGTOP-3615
for Hive 3.x. In HIVE-25795 there seem to be no
indication that upstream wants to fix branch-2.3 too,
so this change should be enough for the moment to secure
Hive users of Bigtop 1.5.

Caveats:
- The Hive branch-2.3 is now at version 2.3.10 (in development),
that is far from our 2.3.6. I don't think it is needed to bump
Hive to 2.3.9/10 at the moment, but it can be discussed.