BIGTOP-3614: allow to change the docker-compose.yml file in the provisioner (#851)
This is a simple change to allow docker-compose to use a different
config file when creating instances. This is useful when testing
new settings and also to keep multiple versions of docker-compose.yml
(if the same version cannot be re-used across multiple OS etc..).

For example, from my recent tests the actual docker-compose.yml
config seems not to work with cgroup v2 and systemd on Debian 10/11.

Add also a simple configuration file called docker-compose-cgroupsv2.yml
to be used and tested on different OS(s).