BIGTOP-3701: Remove deprecated Hadoop fs-image after uninstalling HDFS-namenode
When uninstalled Hadoop-2.10 and reinstalled Hadoop-3.2,
it would fail to format namenode and The user have to restart
NameNode with the "-rollingUpgrade started" option if a rolling
upgrade is already started; or restart NameNode with the
"-upgrade" option to start a new upgrade. It would confuse some
new Hadoop users.
It's better to remove deprecated Hadoop fs-image file after
uninstalling HDFS-namenode in RPM spec.

Change-Id: I593b1c3cde208605cdb770f4552a8bc423fc68ae
Signed-off-by: Yuqi Gu <yuqi.gu@arm.com>