BIGTOP-3812: Building bigtop-ambari-mpack fails on Debian/Ubuntu due to unwanted binary files
Building bigtop-ambari-mpack fails on Debian/Ubuntu due to unwanted binary files.
We will move 'dev-support' to Bigtop provisioner and rename it to bigtop-stack-provisioner.
It is used for Mpack provisioner which is the test tool for Mpack developers.
After 3.2 release, we it will be used for Ambari UI test framework and Ambair-bigtop-demo.

Change-Id: Ibe310d7ab979c055eacfc8e14daa27ca6aa415e3
Signed-off-by: Yuqi Gu <yuqi.gu@arm.com>